Look for build regression in the classify command
